Consolidate multiple exception handlers
An earlier one was more lenient with network errors, which can sometimes
happen with updater. A newer handler overrode the existing one and
failed on everything.

This combines the two so we handle network errors accordingly and
properly fail on others.

We still run the exception handler super early to catch any issues
during the bootup process.

See #122