Fix: Extensions loader in webpack
This fixes the error that comes up when using a newer version of
wp-calypso. It requires a custom webpack loader to translate the
extensions reducers into a js module.

This also pulls up the calypso submodule to the point that the
extensions loader PR was merged there.

See: https://github.com/Automattic/wp-calypso/pull/11728