Upgrade Babel to 7.0.0 beta 53
This patch upgrades Babel to 7.0.0 beta 53 and adds a `babel.config.js` file.

The config file uses Babel Env Preset to specify a compilation target (Electron) and
is used to compile the `wp-desktop` files and also the `calypso/server` ones that are
included by `desktop/server/server.js`.

With `babel.config.js`, Babel will no longer load `calypso/.babelrc.js`, so the desired
config needs to be specified by the parent project.

This patch is a prerequisite to upgrading Babel in Calypso, too.