Fix how we look for modules in the test build (#589)
We were missing a relative 'node_modules' as the first entry, which forced modules to always be found in one the latter places in the resolve.modules config. This led to libraries loading the wrong version of some dependencies, depending on what was in the root, or not finding nested dependencies.

Also, set the mode on the webpack build for the test resources to suppress the associated warning.