Types: Add basic TypeScript support
Calypso itself has adopted TypeScript for gradually typing th
application. It uses a Babel plugin to strip out type annotions
before hitting the rest of the existing build pipeline.

This patch is intended to allow for using the new TypeSript-based
code within th desktop application, which currently is breaking
because of it.