Load SVG images (like Gridicons) with webpack file-loader
Use the loader to generate correct URLs for server-side rendered images and icons,
e.g., in `EnvironmentBadge`. Emitting the files is not necessary, as the actual
assets are provided by the client Calypso bundle.

The desktop server bundle is every used only in production mode. Development server
runs the classic Calypso as a server running on `calypso.localhost:3000` and points
the Electron browser to that server.