[deps] Replace prebuilt spellchecker with source
The prebuilt repo in use for this module (maintained by SSBC) has not been
updated since December 2017 (as of this writing), and available prebuilt
binaries do not support versions of Electron ≥ 2.0.0. The intention of the
original developers in using the prebuilt binaries was to facilitate building
for all three platforms (Windows, Mac and Linux) on Mac only. However, the
prebuilt binary repository has not been updated to keep up with newer
versions of Electron. Therefore, we should proceed to build each platform
within a native environment using the original source for the spellchecker
module.

Because at least one of the native sub-dependencies of this package has a
post-install script that will attempt to compile against the host/Calypso
node environment, we amend the .npmrc to force auto-compilation against
the version of Electron in use in order to avoid build errors.