[tests] Fix e2e tests
This commit fixes addresses breakage of the e2e test suite as a result of upgrading
Chromedriver and Electron.

The root cause of the breakage was host environment contention, where the
wp-desktop boot script was unable to "see" Electron in the immediate host
environment when booted by Chromedriver ("Electron is not defined" error).

By decoupling the initialization of Chromedriver from wp-desktop, contention
between the two processes is minimized, thereby allowing the test suite to
perform more reliably.

In addition, e2e test logging is added to the Circle config to faciliate more effective
debugging going forward.