[Electron] Upgrade To v4.2.9 (#757)
# Key Changes

### Dependency Updates

* Upgrade electron, electron-builder and electron-updater

* Update electron-chromedriver

* Add rxjs (for spellchecker fix)

* Replace prebuilt spellchecker with source, upgrade to electron-spellchecker@2.2.1

The prebuilt repo in use for the electron-spellchecker module (maintained by SSBC) has not been updated since December 2017 (as of this writing), and available prebuilt binaries do not support versions of Electron ≥ 2.0.0. The intention of the original developers in using the prebuilt binaries was to facilitate building for all three platforms (Windows, Mac and Linux) on Mac only. However, the prebuilt binary repository has not been updated to keep up with newer versions of Electron. Therefore, we should proceed to build each platform within a native environment using the original source for the spellchecker module.

### Fix Breaking API Changes

* Deprecated: makeSingleInstance

* Updated: crashReporter

* session.setProxy (undocumented in official Breaking Changes docs)

### CI

* Update Xcode image to 10.1.0

* Add libxext-dev and libx11-dev for native Linux compilation (required by electron-spellchecker)

* Enable native Windows build in Circle

### Tests

* Fix e2e tests

This change addresses breakage of the e2e test suite as a result of upgrading Chromedriver and Electron.

The root cause of the breakage was host environment contention, where the wp-desktop boot script was unable to "see" Electron in the immediate host environment when booted by Chromedriver ("Electron is not defined" error).

By decoupling the initialization of Chromedriver from wp-desktop, contention between the two processes is minimized, thereby allowing the test suite to perform more reliably.

In addition, e2e test logging is added to the Circle config to facilitate more effective debugging going forward.

* Update user-agent info

### Enable Native Builds on Windows

- Amends the .npmrc to ensure native modules with postinstall script build against the Electron runtime.

- Adds a workaround for the buggy release of node-gyp included in the current
version of nodejs being used by Calypso. This workaround unblocks Windows builds
until Calypso node includes node-gyp >= 6.0.1, at which point this workaround
can be reverted. To use, run:

powershell:

> npm config set node_gyp "$(Get-Location)\node_modules\node-gyp\bin\node-gyp.js"

cmd:

> npm config set node_gyp "%cd%\node_modules\node-gyp\bin\node-gyp.js"


### Spellchecker Fixes

Fix regressions in spellchecker behavior by:

(1) Ensuring the spellchecker is attached after the editor iframe has loaded, and

  Ref: https://github.com/electron/electron/issues/21324

(2) Using a dummy observable to circumvent the spellchecker's early exit condition

  Ref: https://github.com/electron-userland/electron-spellchecker/blob/7227d61415c47dfa247828da3b62bbb3330095aa/src/spell-check-handler.js#L206

### Other

* Remove custom notarization patches (unnecessary with updated electron-builder)