[tests] Add end-to-end testing utility (#778)
To-date, executing the end-to-end ("e2e") test suite locally (for debugging or augmentation purposes) has been cumbersome as it involved manually invoking the application, Chromedriver and mocha processes for each test run (as well as shutting down and re-starting each process for subsequent test runs). In addition, developers had to cross-reference the Circle CI script and dashboard to verify the environment variables necessary to invoke the test-suite.

This commit adds a utility script that: 

1. makes the required environment variables explicit 
2. automatically invokes all requisite processes, and 
3. automatically kills invoked processes upon test completion or error. In addition, this logic ensures a clean exit if the developer interrupts a test run via `SIGINT` ( `CTRL` + `C`) 

These changes combined make running (and re-running) the tests much more streamlined, thereby significantly improving the developer workflow.