ci: Add step to validate tag matches version in package.json
This step runs before the build one, and if the build started from
a tag, checks that the version in package.json matches the tag.

I'm not happy that we now have to run this step on every build, but it
also adds less that 15s to a build that is in the order of minutes, so
it seems like a reasonable tradeoff to prevent future errors when
tagging new versions.